TranscriptionAnalyzer/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
llama-cpp-python==0.2.11
sentence-transformers==2.2.2
torch>=2.0.1 